Move the validation providers to the test utilities package such that they can be reused.
